pyproject.toml
[project]
name = "calce-dev"
version = "0.0.0"
description = "Development environment for Calce"
requires-python = ">=3.12"

[dependency-groups]
dev = [
    "maturin>=1.0,<2.0",
    "pytest>=8.0",
    "ruff>=0.9",
    "anthropic>=0.45",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9",
]

[tool.ruff]
target-version = "py312"
line-length = 120
extend-exclude = ["services/calce-db/alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
